+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Hotfix]
+### Changed
+- i18n locales are now enabled with `ENABLED_LOCALES` env variable
+
 ## [0.3.0] - 2018-05-03
 ### Added
 - Node navbar: to accompany all node pages